Kernel/Graphics: Move Bochs graphics related code into a separate folder
