+++ Kernel/Graphics/Bochs/GraphicsAdapter.cpp
@@ -10,7 +10,7 @@
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>
-#include <Kernel/Graphics/BochsGraphicsAdapter.h>
+#include <Kernel/Graphics/Bochs/GraphicsAdapter.h>
 #include <Kernel/Graphics/Console/ContiguousFramebufferConsole.h>
 #include <Kernel/Graphics/GraphicsManagement.h>
 #include <Kernel/IO.h>
