+++ Kernel/Graphics/Bochs/GraphicsAdapter.h
