+++ Kernel/Graphics/GraphicsManagement.cpp
@@ -7,7 +7,7 @@
 #include <AK/Singleton.h>
 #include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/CommandLine.h>
-#include <Kernel/Graphics/BochsGraphicsAdapter.h>
+#include <Kernel/Graphics/Bochs/GraphicsAdapter.h>
 #include <Kernel/Graphics/GraphicsManagement.h>
 #include <Kernel/Graphics/Intel/NativeGraphicsAdapter.h>
 #include <Kernel/Graphics/VGACompatibleAdapter.h>
